refactor: move ValidateUpgradeFields to upgrade.go and make it private
